Fix parsing of aliased stream records * This marks first commit where all Hitman: Blood Money exporter issues should be gone

+++ src/G1AT/Hitman4Dialog.hpp
@@ -193,7 +193,7 @@ struct Hitman4STRFile
 
   bool Save(const StringView8CI &savePath);
 
-  OrderedMap<uint32_t, Hitman4WAVRecord> recordMap;
+  OrderedMap<uint64_t, Hitman4WAVRecord> recordMap;
   std::list<std::vector<char>> extraData;
 
   Hitman4STRHeader                    header;
